# go-event-ticket-service/cmd/swag/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API Document for Event Ticket Service",
        "title": "API Document for Event Ticket Service",
        "termsOfService": "https://github.com/nguyenphamhoangvu852004/ticket4u",
        "contact": {
            "name": "nguyenphamhoangvu",
            "url": "https://github.com/nguyenphamhoangvu852004/ticket4u",
            "email": "nguyenphamhoangvu852004@gmail.com"
        },
        "version": "1.0"
    },
    "host": "localhost:8085",
    "basePath": "/api/v1/2025",
    "paths": {
...
        "/eventTimes/{eventTimeId}": {
            "get": {
                "description": "Get details of an event time by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "EventTimes"
                ],
                "summary": "Get event time by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Event Time ID",
                        "name": "eventTimeId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.GetEventTimeByIdRes"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.APIResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Soft delete an event time and its associated tickets",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "EventTimes"
                ],
                "summary": "Soft delete event time",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Event Time ID",
                        "name": "eventTimeId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Delete request body",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.SoftDeleteEventTimeReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.SoftDeleteEventTimeRes"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.APIResponse"
                        }
                    }
                }
            }
        },
...
        "/events/{eventId}/eventTimes": {
            "post": {
                "description": "Create a new event time for a specific event",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "EventTimes"
                ],
                "summary": "Create event time",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Event ID",
                        "name": "eventId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Create event time request",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateEventTimeReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.CreateEventTimeRes"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.APIResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.APIResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.CreateEventTimeReq": {
            "type": "object",
            "properties": {
                "creatorId": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "endTime": {
                    "type": "string"
                },
                "eventId": {
                    "type": "string"
                },
                "startTime": {
                    "type": "string"
                }
            }
        },
        "dto.CreateEventTimeRes": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                }
            }
        },
...
        "dto.GetEventTimeByIdRes": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "endTime": {
                    "type": "string"
                },
                "eventId": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "startTime": {
                    "type": "string"
                }
            }
        },
...
        "dto.SoftDeleteEventTimeReq": {
            "type": "object",
            "required": [
                "deletorId"
            ],
            "properties": {
                "deletorId": {
                    "type": "string"
                },
                "eventTimeId": {
                    "type": "string"
                }
            }
        },
        "dto.SoftDeleteEventTimeRes": {
            "type": "object",
            "required": [
                "deletedAt",
                "eventTimeId"
            ],
            "properties": {
                "deletedAt": {
                    "type": "integer"
                },
                "eventTimeId": {
                    "type": "string"
                }
            }
        },
...
        "response.APIResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "description": "data return null not show"
                },
                "error": {
                    "description": "Error return null not show"
                },
                "message": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "Bearer": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}